added events

+++ src/pages/sign-in/index.js
@@ -1,7 +1,13 @@
 import SignInComponent from "@/components/login/sign-in"
 
 export default function SignIn() {
-  sessionStorage.clear()
+  if (typeof window !== 'undefined') {
+    const protectedPage = sessionStorage.getItem('protectedPage')
+    if (protectedPage) {
+      sessionStorage.clear()
+    }
+  }
+
   return (
     <SignInComponent />
   )
